pyproject.toml

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-search"
version = "0.1.0"
requires-python = ">=3.10"

# Local-path dependency for zip-based testing.
# Replace with git URL in real usage:
# common-adapters @ git+ssh://git@github.com/YOUR_ORG/common-adapters.git@v1.0.0
dependencies = [
  "common-adapters @ git+https://github.com/Coforge-forgeX/forgexpackages.git@v1.0.0",

  # MCP server stack
  "fastmcp>=2.9.1",
  "fastapi-mcp>=0.3.4",
  "mcp[cli]>=1.10.0",
  "uvicorn>=0.24.0",
  "starlette>=0.27.0",

  # Auth & config
  "python-dotenv",
  "PyJWT>=2.8.0",
  "passlib[argon2]>=1.7.4",
  "argon2-cffi>=21.3.0",

  # Database / storage
  "pymongo>=4.15.0",
  "psycopg2-binary",
  "asyncpg>=0.30.0",
  "sqlalchemy>=2.0.0",
  "certifi",

  # AI / LLM
  "openai>=1.0.0",
  "langchain",
  "langchain-community",
  "langchain-openai",
  "langchain-core",
  "langchain-text-splitters",
  "langchain-ollama",
  "langchain-chroma",

  # RAG & knowledge graph
  "lightrag-hku>=1.3.9",
  "neo4j>=5.28.1",
  "faiss-cpu>=1.11.0",
  "sentence-transformers>=5.0.0",
  "chromadb",

  # Document processing
  "PyMuPDF",
  "pypdf",
  "pypdf2>=3.0.1",
  "python-docx>=1.2.0",
  "pytesseract",
  "easyocr>=1.7.2",
  "pdf2image>=1.17.0",
  "unstructured==0.7.12",
  "azure-ai-documentintelligence>=1.0.0",

  # Web crawling
  "crawl4ai",
  "aiohttp>=3.8.0",

  # Cloud
  "azure-storage-blob>=12.26.0",
  "boto3>=1.28.0",

  # Utilities
  "numpy>=1.24.0",
  "tiktoken",
  "requests",
  "nest-asyncio>=1.6.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
kb-server = "agent_search.server.main:http_app"
